refactor: reorganize dataclass definitions and improve modularity
- Restructured dataclass modules by introducing a `core` and `ros_msgs` package to better separate general and ROS-specific implementations.
- Updated imports across the project to reflect the new hierarchy, ensuring consistency and functionality.
- Refined documentation and examples with the updated namespaces for improved clarity and usability.

Issue TCT-62
