+++ documentation/rosbag_usage.md
@@ -1,3 +1,5 @@
+from trajectory_container_tools.dataclasses import NavMsgsOdometry
+
 # ROS Bag To TCT  Usage Guide
 
 This guide covers how to extract trajectory data from ROS bag files using Trajectory Container Tools (TCT).
@@ -115,6 +117,7 @@ features_config = {
 ### Example 1: Multi-Topic Robot Data
 
 ```python
+import trajectory_container_tools.dataclasses.ros_msgs.ros2_stamped_dataclass
 from pathlib import Path
 import trajectory_container_tools as tct
 
@@ -123,7 +126,7 @@ rosbag_path = Path("experiments/robot_nav_2024_01_15.db3").parent
 # Configure multiple ROS topics
 features_config = {
         # Navigation data
-        '/odometry':          tct.dataclasses.NavMsgsOdometry,
+        '/odometry':          trajectory_container_tools.dataclasses.ros_msgs.ros2_stamped_dataclass.NavMsgsOdometry,
         '/ground_truth_pose': ('GroundTruth', 'pose_pose_position_x', 'pose_pose_position_y',
                                'pose_pose_orientation_z'),
 
@@ -189,18 +192,19 @@ angular_vel = tc_odom.topic_odom.twist.angular
 ### Example 3: Time-Sliced Extraction
 
 ```python
+import trajectory_container_tools.dataclasses.ros_msgs.ros2_stamped_dataclass
 import trajectory_container_tools as tct
 
 # Extract only a portion of the rosbag
 partial_data = tct.from_rosbag(
-    rosbag_path=rosbag_path,
-    dataset_info="Partial trajectory - middle section",
-    features_config={
-        '/odometry': tct.dataclasses.NavMsgsOdometry
-        },
-    start=1000,  # Start from message index 1000
-    stop=5000    # Stop at message index 5000
-)
+        rosbag_path=rosbag_path,
+        dataset_info="Partial trajectory - middle section",
+        features_config={
+                '/odometry': trajectory_container_tools.dataclasses.ros_msgs.ros2_stamped_dataclass.NavMsgsOdometry
+                },
+        start=1000,  # Start from message index 1000
+        stop=5000  # Stop at message index 5000
+        )
 
 print(f"Partial trajectory length: {partial_data.topic_odometry.trajectory_len}")
 ```
@@ -210,6 +214,7 @@ print(f"Partial trajectory length: {partial_data.topic_odometry.trajectory_len}"
 ### Registering Custom Messages
 
 ```python
+import trajectory_container_tools.dataclasses.ros_msgs.ros2_stamped_dataclass
 import trajectory_container_tools as tct
 from rosbags.typesys import get_types_from_msg
 
@@ -226,20 +231,20 @@ uint32 timestamp
 typestore = tct.ros.get_rosbag_typestore_auto_distro()
 
 typestore.register(
-            get_types_from_msg(custom_msg_def, 'custom_msgs/msg/TrajectoryPoint')
+        get_types_from_msg(custom_msg_def, 'custom_msgs/msg/TrajectoryPoint')
         )
 
 typestore = tct.ros.register_non_native_msgs(typestore)
 
 tc_with_custom_type = tct.from_rosbag(
-    rosbag_path=rosbag_path,
-    dataset_info="Trajectory with custom type",
-    features_config={
-        'trajectory_points': ('TrajectoryPoint', 'x', 'y', 'theta', 'velocity')
-        '/odometry': tct.dataclasses.NavMsgsOdometry
-        },
-    typestore=typestore
-)
+        rosbag_path=rosbag_path,
+        dataset_info="Trajectory with custom type",
+        features_config={
+                'trajectory_points': ('TrajectoryPoint', 'x', 'y', 'theta', 'velocity')
+                '/odometry':         trajectory_container_tools.dataclasses.ros_msgs.ros2_stamped_dataclass.NavMsgsOdometry
+                },
+        typestore=typestore
+        )
 ```
 
 ### Custom Dataclass for Complex Messages
@@ -289,8 +294,10 @@ Error: Topic '/odom' not found in rosbag
 ```
 
 **Solutions:**
+
 ```python
-import trajectory_container_tools as tct 
+import trajectory_container_tools.dataclasses.ros_msgs.ros2_stamped_dataclass
+import trajectory_container_tools as tct
 
 # Check available topics
 tct.check_bag_topics(rosbag_path)
@@ -301,7 +308,8 @@ common_odom_topics = ['/odom', '/odometry', '/robot/odom', '/base_link/odom']
 # Try different topic names
 for topic in common_odom_topics:
     try:
-        container = tct.extract_rosbag_feature(rosbag_path, topic, tct.dataclasses.NavMsgsOdometry)
+        container = tct.extract_rosbag_feature(rosbag_path, topic,
+                                               trajectory_container_tools.dataclasses.ros_msgs.ros2_stamped_dataclass.NavMsgsOdometry)
         print(f"Success with topic: {topic}")
         break
     except:
