+++ src/trajectory_container_tools/dataclasses/core/__init__.py
@@ -1,0 +1,1 @@
+# coding=utf-8
