+++ src/trajectory_container_tools/dataclasses/core/base_trajectory_dataclass.py
