+++ src/trajectory_container_tools/temporal/sequence_ordering.py
@@ -3,12 +3,10 @@ from typing import Any, Union
 
 import numpy as np
 
-from trajectory_container_tools.dataclasses.base_trajectory_dataclass import (
+from trajectory_container_tools.dataclasses.core.base_trajectory_dataclass import (
     NestedBaseTrajectoryDataclass,
 )
-from trajectory_container_tools.dataclasses.ros2_feature_dataclass import (
-    RosStampedDataclass,
-)
+from trajectory_container_tools.dataclasses import RosStampedDataclass
 from trajectory_container_tools.utils.shadow_data_container import (
     fetch_timestamps_from_shadow_data_container,
 )
