+++ src/trajectory_container_tools/utils/containers_sanity_checks.py
@@ -3,7 +3,7 @@ from typing import Optional, Tuple
 
 import numpy as np
 
-from ..dataclasses.abstract_trajectory_dataclass import AbstractTrajectoryDataclass
+from trajectory_container_tools.dataclasses.core.abstract_trajectory_dataclass import AbstractTrajectoryDataclass
 
 
 def containers_timestep_alignment_sanity_check(
