+++ tests/tests_dataclasses/tests_abstract_trj_dataclasses/test_abstract_multifeature_dataclass.py
@@ -9,10 +9,8 @@ import pytest
 from trajectory_container_tools import (
     AbstractMultifeatureDataclass,
     )
-from trajectory_container_tools.dataclasses.ros2_feature_dataclass import (
-    RosStampedDataclass,
-)
-from trajectory_container_tools.dataclasses.ros2_primitive_dataclass import Header
+from trajectory_container_tools.dataclasses import RosStampedDataclass
+from trajectory_container_tools.dataclasses.ros_msgs.ros2_primitive_dataclass import Header
 from trajectory_container_tools.temporal.timestamps import Timestamps
