+++ tests/tests_dataclasses/tests_abstract_trj_dataclasses/test_abstract_trajectory_dataclass.py
@@ -9,7 +9,7 @@ from .mock import (
     MockTrajectoryComposedParent,
     MockTrajectoryComposedParentNestedOnly,
 )
-from trajectory_container_tools.dataclasses.abstract_trajectory_dataclass import (
+from trajectory_container_tools.dataclasses.core.abstract_trajectory_dataclass import (
     AbstractTrajectoryDataclass,
 )
 from trajectory_container_tools.dataclasses.panda_dataframe_feature_dataclass import (
