+++ tests/tests_extractor/test_rosbag_to_trajectory_dataclass.py
@@ -9,18 +9,14 @@ from trajectory_container_tools.extractor.rosbag_to_tct import (
     from_rosbag,
     extract_rosbag_feature,
 )
-from trajectory_container_tools.dataclasses.ros2_primitive_dataclass import Header
+from trajectory_container_tools.dataclasses.ros_msgs.ros2_primitive_dataclass import Header
 
-from trajectory_container_tools.dataclasses.ros2_feature_dataclass import (
-    AckermannMsgsAckermannDrive,
-    AckermannMsgsAckermannDriveStamped,
-    NavMsgsOdometry,
-    RosStampedDataclass,
-    Scan,
-    SensorMsgsImu,
+from trajectory_container_tools.dataclasses.ros_msgs.ros2_non_trajectory_dataclass import (
     Tf2MsgsTFMessage,
-    VescMsgsVescImuStamped,
-)
+    )
+from trajectory_container_tools.dataclasses import AckermannMsgsAckermannDrive, \
+    AckermannMsgsAckermannDriveStamped, NavMsgsOdometry, RosStampedDataclass, Scan, SensorMsgsImu, \
+    VescMsgsVescImuStamped
 from trajectory_container_tools.temporal.timestamps import (
     TimestampCausalOrderingError,
 )
