+++ tests/tests_extractor/test_rosbag_to_trajectory_dataclass_benchmark.py
@@ -10,10 +10,7 @@ from trajectory_container_tools import (
     check_bag_topics,
     extract_rosbag_feature,
 )
-from trajectory_container_tools.dataclasses.ros2_feature_dataclass import (
-    NavMsgsOdometry,
-    RosStampedDataclass,
-)
+from trajectory_container_tools.dataclasses import NavMsgsOdometry, RosStampedDataclass
 
 
 @pytest.fixture(scope="function")
